changelog/requirements.txt
yamale==2.0.1
pyyaml==5.3.1
natsort==7.0.1
requests==2.23.0
